requirements.txt
google-genai>=1.0.0
pydantic>=2.6.0
python-dotenv>=1.0.0
markdown-it-py>=3.0.0
PyYAML>=6.0.0
